requirements.txt
aiogram~=3.24.0
python-dotenv~=1.2.1
sqlalchemy~=2.0.46
pandas~=3.0.0
matplotlib~=3.10.8
seaborn~=0.13.2